[json] Add session 7 lectures and PDF
This commit adds session 7 lectures and a PDF to the data.json file. The
session includes lectures on the equation of a circle (part 1 and part
2) as well as a lecture on the cw&hw of the equation of a circle. The
commit also includes a link to the PDF file for session 7. This update
provides additional learning resources for session 7 students. No issues
are associated with this commit.

[json] Add session 7 lectures and PDF

This commit adds session 7 lectures and a PDF file. Enriches session 7
with lectures on the equation of a circle (part 1 and part 2) and a
lecture on the cw&hw of the equation. Facilitates access to learning
materials by including a link to the PDF file. This update enhances the
resources available to session 7 students. No associated issues.